feat: Moved unit tests to seperate folder
